Changed double to float for free_2df

+++ TKmatrix.h
@@ -42,7 +42,7 @@ void free_uinv(double** uinv);
 int get_blas_rows(double** uinv);
 int get_blas_cols(double** uinv);
 float** malloc_2df(int rows,int cols);
-void free_2df(double** uinv);
+void free_2df(float** uinv);
 #ifdef __cplusplus
 }
 #endif
